update spect config

+++ opengate/geometry/volumes.py
@@ -1123,7 +1123,7 @@ class ImageVolume(VolumeBase, solids.ImageSolid):
             label_to_mu[label] = mu
 
         arr = itk.GetArrayViewFromImage(self.label_image)
-        mu_arr = arr.copy().astype("float32")
+        mu_arr = arr.copy().astype("float")
         for label, mu in label_to_mu.items():
             mu_arr[mu_arr == label] = mu
         itk_mu_img = itk.GetImageFromArray(mu_arr)
